Moved logic out of component and into kea
